[MIG] l10n_ar_afipws: Migration to 19.0
