--- l10n_ar_afipws/data/ir.actions.url_data.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="action_afip_padron" model="ir.actions.act_url">
-        <field name="name">AFIP - Consulta Constancia Padron</field>
-        <field name="url">https://seti.afip.gob.ar/padron-puc-constancia-internet/ConsultaConstanciaAction.do</field>
-        <field name="target">new</field>
-    </record>
-</odoo>
