--- l10n_ar_afipws/security/ir.model.access.csv
@@ -1,11 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_afipws_connection_manager,afipws.connection.manager,model_afipws_connection,base.group_user,1,1,1,1
-access_afipws_connection_user,afipws.connection.user,model_afipws_connection,,1,0,0,0
-access_afipws_certificate_alias_manager,afipws.certificate.alias.manager,model_afipws_certificate_alias,base.group_system,1,1,1,1
-access_afipws_certificate_alias_user,afipws.certificate.alias.user,model_afipws_certificate_alias,base.group_user,1,0,0,0
-access_afipws_certificate_manager,afipws.certificate.manager,model_afipws_certificate,base.group_system,1,1,1,1
-access_afipws_certificate_user,afipws.certificate.user,model_afipws_certificate,base.group_user,1,0,0,0
-access_a5_wizard,res.partner.update.from.padron.wizard,model_res_partner_update_from_padron_wizard,base.group_user,1,1,1,1
-access_afipws_upload_certificate_wizard,afipws.upload_certificate.wizard,model_afipws_upload_certificate_wizard,base.group_system,1,1,1,1
-access_a5_field,res.partner.update.from.padron.field,model_res_partner_update_from_padron_field,base.group_user,1,1,1,1
-access_res_partner_update_from_padron_field,res.partner.update.from.padron.field,model_res_partner_update_from_padron_field,base.group_user,1,0,0,0
