--- l10n_ar_afipws/security/security.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo noupdate="1">
-    <record id="afipws_certificate_alias_multi_company_rule" model="ir.rule">
-            <field name="name">Afipws Certificate Alias Multi-company</field>
-            <field name="model_id" ref="l10n_ar_afipws.model_afipws_certificate_alias"/>
-            <field eval="True" name="perm_read"/>
-            <field eval="True" name="perm_write"/>
-            <field eval="True" name="perm_create"/>
-            <field eval="True" name="perm_unlink"/>
-            <field name="domain_force">[('company_id', 'in', company_ids)]</field>
-    </record>
-
-    <record id="afipws_certificate_multi_company_rule" model="ir.rule">
-            <field name="name">Afipws Certificate Multi-company</field>
-            <field name="model_id" ref="l10n_ar_afipws.model_afipws_certificate"/>
-            <field eval="True" name="perm_read"/>
-            <field eval="True" name="perm_write"/>
-            <field eval="True" name="perm_create"/>
-            <field eval="True" name="perm_unlink"/>
-            <field name="domain_force">[('alias_id.company_id','child_of',company_ids)]</field>
-    </record>
-</odoo>
