--- l10n_ar_afipws/views/afipws_certificate_alias_view.xml
@@ -1,97 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="view_afipws_certificate_alias_form" model="ir.ui.view">
-        <field name="name">afipws.certificate_alias.form</field>
-        <field name="model">afipws.certificate_alias</field>
-        <field name="arch" type="xml">
-            <form string="AFIP Certificates">
-                <field name="service_type" invisible="1"/>
-
-                <header>
-                    <button name="action_confirm"
-                        type="object"
-                        invisible="state != 'draft'"
-                        string="Confirm"
-                        class="oe_highlight"
-                        help="Confirm and generate Key"/>
-                    <button name="action_create_certificate_request"
-                        type="object"
-                        invisible="state != 'confirmed'"
-                        class="oe_highlight"
-                        string="Create Certificate Request"/>
-                    <button name="action_cancel"
-                        type="object"
-                        invisible="state not in ('draft', 'confirmed')"
-                        string="Cancel"/>
-                    <button name="action_to_draft"
-                        type="object"
-                        invisible="state != 'cancel'"
-                        string="To Draft"/>
-                    <field name="state" widget="statusbar"/>
-                </header>
-                <sheet string="AFIP Certificates">
-                    <group>
-                        <group>
-                            <field name="company_id" readonly="state != 'draft'" groups="base.group_multi_company"/>
-                            <field name="type" readonly="state != 'draft'"/>
-                            <field name="service_type" readonly="state != 'draft'" groups="base.group_no_one"/>
-                            <field name="service_provider_cuit" readonly="state != 'draft'" invisible="service_type != 'outsourced'" required="service_type == 'outsourced'"/>
-                            <field name="company_cuit" readonly="state != 'draft'" invisible="service_type != 'in_house'" required="service_type == 'in_house'"/>
-                        </group>
-                        <group>
-                            <field name="common_name" readonly="state != 'draft'"/>
-                            <field name="country_id" readonly="state != 'draft'"/>
-                            <field name="state_id" readonly="state != 'draft'"/>
-                            <field name="city" readonly="state != 'draft'"/>
-                        </group>
-                    </group>
-                    <notebook>
-                        <page string="Ceritificates">
-                            <field name="certificate_ids"/>
-                        </page>
-                        <page string="Key" groups="base.group_no_one">
-                            <field name="key" readonly="state != 'draft'"/>
-                        </page>
-                    </notebook>
-                </sheet>
-            </form>
-        </field>
-    </record>
-
-    <record id="view_afipws_certificate_alias_filter" model="ir.ui.view">
-        <field name="name">afipws.certificate_alias.filter</field>
-        <field name="model">afipws.certificate_alias</field>
-        <field name="arch" type="xml">
-            <search string="AFIP Certificates">
-                <field name="type"/>
-                <field name="common_name"/>
-                <field name="company_id"/>
-                <field name="state"/>
-            </search>
-        </field>
-    </record>
-
-    <record id="view_afipws_certificate_alias_tree" model="ir.ui.view">
-        <field name="name">afipws.certificate_alias.list</field>
-        <field name="model">afipws.certificate_alias</field>
-        <field name="arch" type="xml">
-            <list string="AFIP Certificates">
-                <field name="type"/>
-                <field name="common_name"/>
-                <field name="company_id"/>
-                <field name="state"/>
-            </list>
-        </field>
-    </record>
-
-    <record model="ir.actions.act_window" id="act_afipws_certificate_alias">
-        <field name="name">AFIP Certificate</field>
-        <field name="type">ir.actions.act_window</field>
-        <field name="res_model">afipws.certificate_alias</field>
-        <field name="view_mode">list,form</field>
-    </record>
-
-    <menuitem name="Certificates" action="act_afipws_certificate_alias" id="menu_action_afipws_certificate_alias" parent="menu_afipws"/>
-
-</odoo>
