--- l10n_ar_afipws/views/afipws_certificate_view.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="view_afipws_certificate_form" model="ir.ui.view">
-        <field name="name">afipws.certificate.form</field>
-        <field name="model">afipws.certificate</field>
-        <field name="arch" type="xml">
-            <form string="AFIP Certificates">
-                <header>
-                    <button name="%(action_upload_certificate)d"
-                        type="action"
-                        invisible="state != 'draft'"
-                        class="oe_highlight"
-                        string="Upload Certificate"/>
-                    <button name="action_confirm"
-                        type="object"
-                        invisible="state != 'draft'"
-                        string="Confirm"/>
-                    <button name="action_to_draft"
-                        type="object"
-                        invisible="state != 'cancel'"
-                        string="To Draft"/>
-                    <button name="action_cancel"
-                        type="object"
-                        invisible="state not in ('draft', 'confirmed')"
-                        string="Cancel"/>
-                    <field name="state" widget="statusbar"/>
-                </header>
-                <sheet string="AFIP Certificates">
-                    <group>
-                        <field name="csr" groups="base.group_no_one" readonly="state != 'draft'"/>
-                        <field name="request_filename" invisible="1"/>
-                        <field name="request_file" readonly="state != 'draft'" filename="request_filename"/>
-                        <field name="crt" groups="base.group_no_one" readonly="state != 'draft'"/>
-                    </group>
-                </sheet>
-            </form>
-        </field>
-    </record>
-
-    <record id="view_afipws_certificate_tree" model="ir.ui.view">
-        <field name="name">afipws.certificate.list</field>
-        <field name="model">afipws.certificate</field>
-        <field name="arch" type="xml">
-            <list string="AFIP Certificates">
-                <field name="create_date"/>
-                <field name="state"/>
-            </list>
-        </field>
-    </record>
-</odoo>
