--- l10n_ar_afipws/views/afipws_menuitem.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <menuitem id="menu_afipws" name="AFIP Web Services" parent="base.menu_custom" sequence="35"/>
-
-</odoo>
