--- l10n_ar_afipws/views/res_config_settings.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-        <record id="res_config_settings_view_form" model="ir.ui.view">
-            <field name="name">res.config.settings.view.form.inherit.account</field>
-            <field name="model">res.config.settings</field>
-            <field eval="40" name="priority"/>
-            <field name="inherit_id" ref="account.res_config_settings_view_form"/>
-            <field name="arch" type="xml">
-                <xpath expr="//block[@id='invoicing_settings']" position="inside">
-                    <setting type="header" string="AFIP WS" help="Configure AFIP web services">
-                        <button class="btn-link" icon="fa-arrow-right" name="%(l10n_ar_afipws.act_afipws_certificate_alias)d" string="List certificates" type="action"/>
-                        <button class="btn-link" icon="fa-arrow-right" name="%(l10n_ar_afipws.act_afipws_auth)d" string="List conections" type="action"/>
-                        <label for="afip_ws_env_type"/>
-                        <field class="font-weight-bold" name="afip_ws_env_type"/>
-                    </setting>
-                </xpath>
-            </field>
-        </record>
-</odoo>
