--- l10n_ar_afipws/views/res_partner.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-    <record id="mipymepartner_form" model="ir.ui.view">
-        <field name="name">mipymepartner.form</field>
-        <field name="model">res.partner</field>
-        <field name="inherit_id" ref="account.view_partner_property_form"/>
-        <field name="arch" type="xml">
-            <group name="accounting_entries" position="after">
-                <group name="miPyme" string="Mi pyme">
-                    <field name="mipyme_required"/>
-                    <field name="mipyme_from_amount" invisible="not mipyme_required"/>
-                    <button name="l10n_ar_afipws_fe_min_ammount" string="Check mi Pyme status" type="object"/>
-                </group>
-            </group>
-        </field>
-    </record>
-    <record id="view_partner_form" model="ir.ui.view">
-        <field name="name">res.partner.form</field>
-        <field name="model">res.partner</field>
-        <field eval="90" name="priority"/>
-        <field name="inherit_id" ref="base.view_partner_form"/>
-        <field name="arch" type="xml">
-            <div name="button_box" position="inside">
-                <button class="oe_stat_button" icon="fa-rss" name="%(action_partner_update)d" type="action">
-                    <div class="o_form_field o_stat_info">
-                        <span class="o_stat_value">AFIP A5</span>
-                        <span class="o_stat_text"><field name="last_update_census" readonly="True"/></span>
-                    </div>
-                </button>
-            </div>
-        </field>
-    </record>
-</odoo>
