--- l10n_ar_afipws/wizard/res_partner_update_from_padron_wizard_view.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-    <record id="action_base_partner_update_from_padron" model="ir.actions.act_window">
-        <field name="name">Update Partners From Padron AFIP</field>
-        <field name="res_model">res.partner.update.from.padron.wizard</field>
-        <field name="view_mode">form</field>
-        <field name="target">new</field>
-    </record>
-    <menuitem action="action_base_partner_update_from_padron" groups="base.group_system" id="menu_base_partner_update_from_padron" parent="contacts.res_partner_menu_config"/>
-    <record id="view_base_partner_update_from_padron_form" model="ir.ui.view">
-        <field name="name">res.partner.update.from.padron.wizard.form</field>
-        <field name="model">res.partner.update.from.padron.wizard</field>
-        <field name="arch" type="xml">
-            <form string="Update From Padron AFIP">
-                <field invisible="1" name="state"/>
-                <header/>
-                <sheet>
-                    <group invisible="state != 'finished'" col="1">
-                        <h2>There are no more partners to update for this request...</h2>
-                    </group>
-                    <p invisible="state != 'option'" class="oe_grey">Only Partners with cuit are going to be updated.
-                        <br/>
-                        Select the list of fields you want to update.</p>
-                    <group invisible="state not in ('option',)">
-                        <field name="field_to_update_ids" options="{'no_create': True}" widget="many2many_tags"/>
-                        <field name="update_constancia"/>
-                        <field name="title_case"/>
-                    </group>
-                    <group invisible="state in ('option', 'finished')" col="1">
-                        <h1>
-                            <field required="state == 'selection'" name="partner_id" options="{'no_open': True}"/>
-                        </h1>
-                        <span>
-                            <field name="field_ids" nolabel="1">
-                                <list create="false" editable="top">
-                                    <field name="field"/>
-                                    <field readonly="True" name="old_value"/>
-                                    <field name="new_value"/>
-                                </list>
-                            </field>
-                        </span>
-                    </group>
-                </sheet>
-                <footer>
-                    <button invisible="state in ('option', 'finished')" class="oe_highlight" name="update_selection" string="Update Selection" type="object"/>
-                    <button invisible="state != 'selection'" class="oe_link" name="next_cb" string="Skip these partner" type="object"/>
-                    <button invisible="state != 'option'" class="oe_highlight" name="start_process_cb" string="Update with Manual Check" type="object"/>
-                    <button invisible="state != 'option'" class="oe_highlight" confirm="Are you sure to execute the automatic update of your partners ?" name="automatic_process_cb" string="Update Automatically" type="object"/>
-                    <span invisible="state == 'finished'" class="or_cancel">or
-                        <button class="oe_link oe_inline" special="cancel" string="Cancel"/></span>
-                    <span invisible="state != 'finished'" class="or_cancel">
-                        <button class="oe_link oe_inline" special="cancel" string="Close"/>
-                    </span>
-                </footer>
-            </form>
-        </field>
-    </record>
-    <record id="action_partner_update" model="ir.actions.act_window">
-        <field name="name">Automatic Update from Padron</field>
-        <field name="res_model">res.partner.update.from.padron.wizard</field>
-        <field name="target">new</field>
-        <field name="view_mode">form</field>
-    </record>
-</odoo>
