+++ l10n_ar_fiscal_ws/README.rst
@@ -1,0 +1,101 @@
+.. |company| replace:: ADHOC SA
+
+.. |company_logo| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-logo.png
+   :alt: ADHOC SA
+   :target: https://www.adhoc.com.ar
+
+.. |icon| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-icon.png
+
+.. image:: https://img.shields.io/badge/license-AGPL--3-blue.png
+   :target: https://www.gnu.org/licenses/agpl
+   :alt: License: AGPL-3
+
+=========================================
+Modulo Base para los Web Services de ARCA
+=========================================
+
+Homologation / production:
+--------------------------
+
+First it search for a paramter "afip.ws.env.type" if exists and:
+
+* is production --> production
+* is homologation --> homologation
+
+Else
+
+Search for 'server_mode' parameter on conf file. If that parameter:
+
+* has a value then we use "homologation",
+* if no parameter, then "production"
+
+Incluye:
+--------
+
+* Wizard para instalar los claves para acceder a las Web Services.
+* API para realizar consultas en la Web Services desde OpenERP.
+
+El módulo l10n_ar_fiscal_ws permite a OpenERP acceder a los servicios del ARCA a
+travésde Web Services. Este módulo es un servicio para administradores y
+programadores, donde podrían configurar el servidor, la autentificación
+y además tendrán acceso a una API genérica en Python para utilizar los
+servicios ARCA.
+
+Para poder ejecutar los tests es necesario cargar la clave privada y el
+certificado al archivo test_key.yml.
+
+Tenga en cuenta que estas claves son personales y pueden traer conflicto
+publicarlas en los repositorios públicos.
+
+Installation
+============
+
+To install this module, you need to:
+
+#. Do this ...
+
+Configuration
+=============
+
+To configure this module, you need to:
+
+#. Go to ...
+
+Usage
+=====
+
+To use this module, you need to:
+
+#. Go to ...
+
+.. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
+   :alt: Try me on Runbot
+   :target: http://runbot.adhoc.com.ar/
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues
+<https://github.com/ingadhoc/odoo-argentina/issues>`_. In case of trouble, please
+check there if your issue has already been reported. If you spotted it first,
+help us smashing it by providing a detailed and welcomed feedback.
+
+Credits
+=======
+
+Images
+------
+
+* |company| |icon|
+
+Contributors
+------------
+
+Maintainer
+----------
+
+|company_logo|
+
+This module is maintained by the |company|.
+
+To contribute to this module, please visit https://www.adhoc.com.ar.
