+++ l10n_ar_fiscal_ws/data/ir.actions.url_data.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record id="action_afip_padron" model="ir.actions.act_url">
+        <field name="name">ARCA - Consulta Constancia Padron</field>
+        <field name="url">https://seti.afip.gob.ar/padron-puc-constancia-internet/ConsultaConstanciaAction.do</field>
+        <field name="target">new</field>
+    </record>
+</odoo>
