+++ l10n_ar_fiscal_ws/models/__init__.py
@@ -1,0 +1,11 @@
+##############################################################################
+# For copyright and license notices, see __manifest__.py file in module root
+# directory
+##############################################################################
+from . import arcaws
+from . import arcaws_certificate_alias
+from . import arcaws_certificate
+from . import arcaws_connection
+from . import res_company
+from . import res_config_settings
+from . import res_partner
