+++ l10n_ar_fiscal_ws/security/security.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo noupdate="1">
+    <record id="arcaws_certificate_alias_multi_company_rule" model="ir.rule">
+            <field name="name">Afipws Certificate Alias Multi-company</field>
+            <field name="model_id" ref="l10n_ar_fiscal_ws.model_arcaws_certificate_alias"/>
+            <field eval="True" name="perm_read"/>
+            <field eval="True" name="perm_write"/>
+            <field eval="True" name="perm_create"/>
+            <field eval="True" name="perm_unlink"/>
+            <field name="domain_force">[('company_id', 'in', company_ids)]</field>
+    </record>
+
+    <record id="arcaws_certificate_multi_company_rule" model="ir.rule">
+            <field name="name">Afipws Certificate Multi-company</field>
+            <field name="model_id" ref="l10n_ar_fiscal_ws.model_arcaws_certificate"/>
+            <field eval="True" name="perm_read"/>
+            <field eval="True" name="perm_write"/>
+            <field eval="True" name="perm_create"/>
+            <field eval="True" name="perm_unlink"/>
+            <field name="domain_force">[('alias_id.company_id','child_of',company_ids)]</field>
+    </record>
+</odoo>
