+++ l10n_ar_fiscal_ws/views/arcaws_menuitem.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <menuitem id="menu_arcaws" name="ARCA Web Services" parent="base.menu_custom" sequence="35"/>
+
+</odoo>
