+++ l10n_ar_fiscal_ws/wizard/upload_certificate_view.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record id="view_upload_certificate_form" model="ir.ui.view">
+        <field name="name">arcaws.upload_certificate.wizard.form</field>
+        <field name="model">arcaws.upload_certificate.wizard</field>
+        <field name="arch" type="xml">
+            <form string="Upload Certificate">
+                <group>
+                    <field name="certificate_id" invisible="1"/>
+                    <field name="certificate_file"/>
+                </group>
+                <footer>
+                    <button name="action_confirm" string="Confirm" type="object" class="oe_highlight"/>
+                    or
+                    <button string="Cancel" class="oe_link" special="cancel"/>
+                </footer>
+            </form>
+        </field>
+    </record>
+
+    <record id="action_upload_certificate" model="ir.actions.act_window">
+        <field name="name">ARCA WS - Upload Certificate</field>
+        <field name="type">ir.actions.act_window</field>
+        <field name="res_model">arcaws.upload_certificate.wizard</field>
+        <field name="view_id" ref="view_upload_certificate_form"/>
+        <field name="view_mode">form</field>
+        <field name="target">new</field>
+    </record>
+</odoo>
