--- requirements.txt
@@ -1,11 +1,0 @@
-pyOpenSSL
-M2Crypto
-httplib2>=0.7
-pysimplesoap~=1.8.22
-# fpdf>=1.7.2
-# dbf>=0.88.019
-# Pillow>=2.0.0
-# git+https://github.com/OCA/openupgradelib/@master
-# git+https://github.com/reingart/pyafipws@py3k
-# git+https://github.com/ingadhoc/pyafipws@py3k
-git+https://github.com/filoquin/pyafipws.git@py3k
